Remove unnecessary file and fix type of score

+++ src/Leaderboard.Domain/ValueObjects/PlayerScoreDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace Leaderboard.Domain.ValueObjects
 {
-    public record PlayerScoreDto(Guid PlayerId, string PlayerName, long Score, int PlayerRank);
+    public record PlayerScoreDto(Guid PlayerId, string PlayerName, int Score, int PlayerRank);
 }
